requirements.txt
geopandas>=0.14
pandas>=2.2.2
plotly>=5.22
psycopg2-binary>=2.9.9
pyarrow==16.1.0
python-dotenv>=1.0.1
PyYAML>=6.0.1
requests>=2.32.0
sqlalchemy>=2.0.30
streamlit==1.35.0
tqdm>=4.66.4